feat(cmd): cmd reorg
